+++ src/main/java/absolutelyaya/ultracraft/config/ConfigEntry.java
@@ -52,4 +52,9 @@ public abstract class ConfigEntry<T>
 		this.translationKey = key;
 		return this;
 	}
+	
+	public boolean isDefault()
+	{
+		return value == null || value.equals(defaultValue);
+	}
 }