+++ src/main/java/absolutelyaya/ultracraft/config/ServerConfig.java
@@ -23,7 +23,7 @@ public class ServerConfig extends Config
 	public final IntegerEntry hellObserverInterval = new IntegerEntry("HellObserverInterval", 5);
 	public final BooleanEntry bloodSaturation = new BooleanEntry("BloodSaturation", false);
 	public final BooleanEntry dodgeableOverpump = new BooleanEntry("DodgeableOverpump", false);
-	public final BooleanEntry customLevelsUnlocked = new BooleanEntry("UnlockAllCustomLevels", true);
+	public final BooleanEntry customLevelsUnlocked = new BooleanEntry("UnlockAllCustomLevels", false);
 	public final FloatEntry parryRange = (FloatEntry)new FloatEntry("ParryRange", 3f).setRange(0f, Float.MAX_VALUE);
 	public final FloatEntry coinPunchRange = (FloatEntry)new FloatEntry("CoinPunchRange", 4f).setRange(0f, Float.MAX_VALUE);
 	public final BooleanEntry disableModificationSuppression = new BooleanEntry("DisableModificationSuppression", false);
@@ -35,6 +35,7 @@ public class ServerConfig extends Config
 	public final FloatEntry nailgunDamage = (FloatEntry)new FloatEntry("NailgunDamage", 1f).setRange(0f, Float.MAX_VALUE);
 	//Debug
 	public final BooleanEntry disableFixedStructures = new BooleanEntry("DisableFixedStructures", false);
+	public final IntegerEntry version = new IntegerEntry("ConfigVersion", 0);
 	
 	public ServerConfig(MinecraftServer server)
 	{
@@ -74,6 +75,7 @@ public class ServerConfig extends Config
 		entries.add(new Comment("           Debug stuff"));
 		entries.add(new Comment(" ## ############################# ##  #"));
 		entries.add(disableFixedStructures);
+		entries.add(version);
 		
 		load(server);
 		INSTANCE = this;
@@ -95,6 +97,25 @@ public class ServerConfig extends Config
 	public void load(MinecraftServer server)
 	{
 		super.load(server);
+		boolean updated = false;
+		int newest = 1;
+		Object version = this.version.value;
+		if(version == null)
+			version = 0;
+		for (int i = (int)version; i < newest; i++) //update settings based on changes over versions
+		{
+			switch(i)
+			{
+				case 0 -> {
+					if(customLevelsUnlocked.getValue()) customLevelsUnlocked.setValue(false);
+				}
+			}
+			updated = true;
+		}
+		this.version.setValue(newest);
+		if(updated)
+			save(server);
+		Ultracraft.LOGGER.info("Ultracraft Server Config has been changed based on Version default value Changes.");
 		Ultracraft.LOGGER.info("Ultracraft Server Config Loaded.");
 	}
 }